refactor: reorganize package into tools/, clients/ subpackages
Separate core broker infra from data-prep tools and client utilities:

- tools/: inspect, generate, schema (onboarding pipeline)
- clients/: tiled_cache, query_manifest (catalog consumers)
- Rename register.py → bulk_register.py with deprecation note
- Delete catalog.py (broken import, inlined into cli.py)
- Trim dead code from config.py and utils.py
- Move test_tiled_cache.py from src/ to tests/

All 75 unit tests pass. No functionality removed.

Closes #38
