+++ src/tiled_catalog_broker/tools/generate.py
