+++ src/tiled_catalog_broker/tools/inspect.py
