+++ src/tiled_catalog_broker/tools/schema/catalog_model.yml
