+++ tests/test_schema.py
@@ -23,7 +23,7 @@ import pytest
 # Add project root to path for package imports
 sys.path.insert(0, str(Path(__file__).parent.parent))
 
-from tiled_catalog_broker.schema import (
+from tiled_catalog_broker.tools.schema import (
     ValidationError,
     get_alias_map,
     get_allowed_values,
